Split the utility functions of enforce-column-name rule
